wecom_ops/requirements.txt
flask==3.0.3
pydantic==2.7.4
python-dotenv==1.0.1
redis==5.0.7
pymysql==1.1.1
mysqlclient==2.2.4
rq==1.15.1
uvicorn==0.30.6
gunicorn==22.0.0
structlog==24.1.0
fakeredis==2.23.2
pytest==8.3.3
